Implement DIP for Storage
